requirements.txt
anyio==4.5.0
APScheduler==3.6.3
attrs==24.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2024.8.30
colour==0.1.5
et-xmlfile==1.1.0
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
inflect==7.4.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
more-itertools==10.5.0
mysql-connector-python==9.0.0
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
pkgutil_resolve_name==1.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2024.2
referencing==0.35.1
rpds-py==0.20.0
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.54
styleframe==4.2
tornado==6.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
zipp==3.20.2
